# modulo_administracion_vehiculos/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Vehículos",
        "description": "Documentación de la API del sistema de administración de vehículos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/vehiculos": {
            "get": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Listar todos los vehículos",
                "operationId": "149b9ba714037a3dc86c8d245646f6fa",
                "responses": {
                    "200": {
                        "description": "Lista de vehículos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vehiculo"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Crear un vehículo",
                "operationId": "04854b3c96cbe81512a35a4340be89d6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "marca",
                                    "modelo",
                                    "placa",
                                    "tipo_id"
                                ],
                                "properties": {
                                    "marca": {
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "modelo": {
                                        "type": "string",
                                        "example": "Hilux"
                                    },
                                    "placa": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "tipo_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehículo creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehiculo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Tipo de vehículo no válido"
                    }
                }
            }
        },
        "/api/vehiculos/{id}": {
            "get": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Obtener un vehículo por ID",
                "operationId": "74d0083b6d8cbc7bd03af2e47b5c4a3b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehículo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "vehiculo": {
                                            "$ref": "#/components/schemas/Vehiculo"
                                        },
                                        "tipo_vehiculo": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehículo no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Actualizar un vehículo",
                "operationId": "8ec2230377ed40f3ce7542ad3163463e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "marca": {
                                        "type": "string",
                                        "example": "Chevrolet"
                                    },
                                    "modelo": {
                                        "type": "string",
                                        "example": "Dmax"
                                    },
                                    "placa": {
                                        "type": "string",
                                        "example": "XYZ789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehículo actualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Eliminar un vehículo",
                "operationId": "76e387e5d46c7b473cd89072b5ac88fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Vehículo eliminado"
                    },
                    "404": {
                        "description": "Vehículo no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Vehiculo": {
                "title": "Vehiculo",
                "required": [
                    "placa",
                    "marca",
                    "modelo",
                    "año",
                    "tipo_id",
                    "estado"
                ],
                "properties": {
                    "id_vehiculo": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "placa": {
                        "type": "string",
                        "example": "ABC1234"
                    },
                    "marca": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "modelo": {
                        "type": "string",
                        "example": "Corolla"
                    },
                    "año": {
                        "type": "integer",
                        "example": 2022
                    },
                    "tipo_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "estado": {
                        "type": "string",
                        "example": "Disponible"
                    }
                },
                "type": "object"
            }
        }
    },
...
}